refactor(android-chat): move thread selector above composer
